# docs/API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: ClearDeed API
  version: 1.0.0
  description: End-to-end verified real estate & investment execution platform
  contact:
    name: ClearDeed Advisors LLP
servers:
  - url: https://api.cleardeed.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
paths:
...
  /profile:
    get:
      summary: Get current user profile
      tags:
        - Profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    post:
      summary: Create/complete user profile after OTP verification
      tags:
        - Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
                city:
                  type: string
                profile_type:
                  type: string
                  enum: ['buyer', 'seller', 'investor']
                budget_range:
                  type: string
                  example: "50-100L"
                net_worth_range:
                  type: string
                referral_mobile_number:
                  type: string
              required:
                - full_name
                - email
                - city
                - profile_type
                - referral_mobile_number
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid referral number or missing fields

    put:
      summary: Update user profile
      tags:
        - Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
                city:
                  type: string
      responses:
        '200':
          description: Profile updated
  /profile/mode-select:
    post:
      summary: Select active role (buyer/seller/investor)
      tags:
        - Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: ['buyer', 'seller', 'investor']
      responses:
        '200':
          description: Role selected

  # ======================== PROPERTIES (BUY MODULE) ========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        mobile_number:
          type: string
        full_name:
          type: string
        email:
          type: string
        city:
          type: string
        profile_type:
          type: string
          enum: ['buyer', 'seller', 'investor']
        is_verified:
          type: boolean
...